Fix long movie titles issue
Issue: If any movie had a title longer than a certain amount (~20) it would be printed within the movie tile on more than 1 line.  When this happens it throws off the alignment of the grid on the row below.

Fix: Only use the first 19 characters of the movie title and add ellipses if the title length exceeds 20 characters.

TODO: It would be nice to find a way to display the full title in some other way.  Ideas:
* Display full title as tooltip on mouseover
* Scroll full title on one line back and forth still showing only enough characters to fit on one line within the tile
* Enlarge the tiles to fit 2 lines of text
* Dynamically increase the size of all tiles when a single one needs to increase its size in order to maintain proper grid alignment (this one seems too invasive)
* Change title display in place from shortened title to full title on mouseover (this one will possible break alignment while viewing the title and could make the grid unstable on mouseovers)